fix: extract createRegistry
