+++ packages/ui/src/lib/create-registry.ts
@@ -1,0 +1,21 @@
+import type { ComponentType } from "react";
+import type { ComponentDefinition } from "./simple-component-catalog.js";
+import type { RendererComponentProps } from "./simple-renderer.js";
+
+export type RegistryFromCatalog<
+  C extends { components: Record<string, ComponentDefinition> },
+> = {
+  [K in keyof C["components"]]: ComponentType<
+    RendererComponentProps<C["components"][K]>
+  >;
+};
+
+/**
+ * Creates a type-safe registry from a catalog and component implementations.
+ * Ensures all catalog components are provided with correct prop types.
+ */
+export function createRegistry<
+  C extends { components: Record<string, ComponentDefinition> },
+>(_catalog: C, components: RegistryFromCatalog<C>) {
+  return components;
+}
