+++ packages/ui/src/lib/simple-renderer.tsx
@@ -6,6 +6,7 @@ import {
 } from "./simple-component-catalog.js";
 import z from "zod";
 import { useKopaiData } from "./use-kopai-data.js";
+import { type RegistryFromCatalog } from "./create-registry.js";
 
 type Catalog = ReturnType<typeof createSimpleCatalog>;
 
@@ -59,24 +60,6 @@ export type RendererComponentProps<CD extends ComponentDefinition> =
           | ({ element: BaseElement<InferProps<P>> } & WithoutData)
           | ({ element: BaseElement<InferProps<P>> } & WithData)
       : never;
-
-type RegistryFromCatalog<
-  C extends { components: Record<string, ComponentDefinition> },
-> = {
-  [K in keyof C["components"]]: ComponentType<
-    RendererComponentProps<C["components"][K]>
-  >;
-};
-
-/**
- * Creates a type-safe registry from a catalog and component implementations.
- * Ensures all catalog components are provided with correct prop types.
- */
-export function createRegistry<
-  C extends { components: Record<string, ComponentDefinition> },
->(_catalog: C, components: RegistryFromCatalog<C>) {
-  return components;
-}
 
 /**
  * Base props (no dataSource)
