+++ packages/ui/src/pages/example.tsx
@@ -1,5 +1,6 @@
 import { dashboardCatalog } from "../lib/catalog.js";
-import { Renderer, createRegistry } from "../lib/simple-renderer.js";
+import { Renderer } from "../lib/simple-renderer.js";
+import { createRegistry } from "../lib/create-registry.js";
 import { Card } from "../components/card.js";
 import { Grid } from "../components/grid.js";
 import { Stack } from "../components/stack.js";
